tests/requirements.txt
pyyaml>6.0.0
rdflib
pytest